Life Expectancy Analysis for the Kenyan Counties
This is a system that analyses the Percentage of a county's Population
that will reach the Life Expectancy Age for Kenya.
It Uses as Parameters analysis of Some of the Common Diseases which are
Contributor to deaths in Kenya.
It also Analyses the Sanitation, Water Treatment, Access to a Medical
Physicians